// src/app/page/master/struktur-nomor/struktur-nomor.interface.ts
import { EventEmitter } from '@angular/core';
export interface StrukturNomor {
  kdDepartemen?: string;
  kdStatusHead?: 0;
  kode?: 0;
  kodeExternal?: string;
  namaExternal?: string;
  namaStatus?: string;
  reportDisplay?: string;
  statusEnabled?: true;
  version?: 0
  noRec?: string;
  label?: any;
  value?: any;
}